Fix on order domain

+++ services/order-service/src/main/java/com/denyaar/orderservice/orderline/OrderLineRequest.java
@@ -6,5 +6,5 @@
 
 package com.denyaar.orderservice.orderline;
 
-public record OrderLineRequest(Integer id, Integer orderId, Integer productId, Integer quantity) {
+public record OrderLineRequest(Integer id, Integer orderId, Integer productId, double quantity) {
 }
